--- src/features/Example AuthByUsername/index.ts
@@ -1,3 +1,0 @@
-/* export { LoginModal } from './ui/LoginModal/LoginModal';
-export { LoginSchema } from './model/types/loginSchema';
- */
