--- src/features/Example AuthByUsername/model/selectors/getLoginError/getLoginError.ts
@@ -1,4 +1,0 @@
-/* import { StateSchema } from "app/providers/StoreProvider"
-
-export const getLoginError = (state: StateSchema) => state?.loginForm?.error
- */
