--- src/features/Example AuthByUsername/model/selectors/getLoginError/getLoginError.ttest.ts
@@ -1,18 +1,0 @@
-/* import { StateSchema } from 'app/providers/StoreProvider';
-import { getLoginError } from './getLoginError';
-
-describe('getLoginError.test', () => {
-    test('should return error', () => {
-        const state: DeepPartial<StateSchema> = {
-            loginForm: {
-                error: 'error',
-            },
-        };
-        expect(getLoginError(state as StateSchema)).toEqual('error');
-    });
-    test('should work with empty state', () => {
-        const state: DeepPartial<StateSchema> = {};
-        expect(getLoginError(state as StateSchema)).toEqual(undefined);
-    });
-});
- */
