--- src/features/Example AuthByUsername/model/selectors/getLoginIsLoading/getLoginIsLoading.ts
@@ -1,4 +1,0 @@
-/* import { StateSchema } from 'app/providers/StoreProvider';
-
-export const getLoginIsLoading = (state: StateSchema) => state?.loginForm?.isLoading || false;
- */
