--- src/features/Example AuthByUsername/model/selectors/getLoginPassword/getLoginPassword.testt.ts
@@ -1,18 +1,0 @@
-/* import { StateSchema } from 'app/providers/StoreProvider';
-import { getLoginPassword } from './getLoginPassword';
-
-describe('getLoginPassword.test', () => {
-    test('should return value', () => {
-        const state: DeepPartial<StateSchema> = {
-            loginForm: {
-                password: '123123',
-            },
-        };
-        expect(getLoginPassword(state as StateSchema)).toEqual('123123');
-    });
-    test('should work with empty state', () => {
-        const state: DeepPartial<StateSchema> = {};
-        expect(getLoginPassword(state as StateSchema)).toEqual('');
-    });
-});
- */
