--- src/features/Example AuthByUsername/model/selectors/getLoginPassword/getLoginPassword.ts
@@ -1,4 +1,0 @@
-/* import { StateSchema } from 'app/providers/StoreProvider';
-
-export const getLoginPassword = (state: StateSchema) => state?.loginForm?.password || '';
- */
