--- src/features/Example AuthByUsername/model/selectors/getLoginUsername/getLoginUsername.testt.ts
@@ -1,18 +1,0 @@
-/* import { StateSchema } from 'app/providers/StoreProvider';
-import { getLoginUsername } from './getLoginUsername';
-
-describe('getLoginUsername.test', () => {
-    test('should return value', () => {
-        const state: DeepPartial<StateSchema> = {
-            loginForm: {
-                username: '123123',
-            },
-        };
-        expect(getLoginUsername(state as StateSchema)).toEqual('123123');
-    });
-    test('should work with empty state', () => {
-        const state: DeepPartial<StateSchema> = {};
-        expect(getLoginUsername(state as StateSchema)).toEqual('');
-    });
-});
- */
