--- src/features/Example AuthByUsername/model/selectors/getLoginUsername/getLoginUsername.ts
@@ -1,4 +1,0 @@
-/* import { StateSchema } from 'app/providers/StoreProvider';
-
-export const getLoginUsername = (state: StateSchema) => state?.loginForm?.username || '';
- */
