--- src/features/Example AuthByUsername/model/services/loginByUsername/loginByUsername.ts
@@ -1,37 +1,0 @@
-/* import { createAsyncThunk } from '@reduxjs/toolkit';
-import { AxiosInstance } from 'axios';
-import { User, userActions } from 'entities/User';
-import { USER_LOCALSTORAGE_KEY } from 'shared/const/localstorage';
-import { ThunkConfig } from 'app/providers/StoreProvider';
-
-interface LoginByUsernameProps {
-    username: string;
-    password: string;
-}
-
-export const loginByUsername = createAsyncThunk<
-    User,
-    LoginByUsernameProps,
-    ThunkConfig<string>
->(
-    'login/loginByUsername',
-    async (authData, thunkApi) => {
-        const { extra, dispatch, rejectWithValue } = thunkApi;
-
-        try {
-            const response = await extra.api.post<User>('/login', authData);
-
-            if (!response.data) {
-                throw new Error();
-            }
-
-            localStorage.setItem(USER_LOCALSTORAGE_KEY, JSON.stringify(response.data));
-            dispatch(userActions.setAuthData(response.data));
-            return response.data;
-        } catch (e) {
-            console.log(e);
-            return rejectWithValue('error');
-        }
-    },
-);
- */
