--- src/features/Example AuthByUsername/model/types/loginSchema.ts
@@ -1,7 +1,0 @@
-/* export interface LoginSchema {
-    username: string;
-    password: string;
-    isLoading: boolean;
-    error?: string;
-}
- */
