--- src/features/Example AuthByUsername/ui/LoginForm/LoginForm.async.ts
@@ -1,9 +1,0 @@
-/* import { FC, lazy } from 'react';
-import { LoginFormProps } from './LoginForm';
-
-export const LoginFormAsync = lazy <FC<LoginFormProps>>(() => new Promise((resolve) => {
-    // @ts-ignore
-    // ТАК В РЕАЛЬНЫХ ПРОЕКТАХ НЕ ДЕЛАТЬ!!!!! ДЕЛАЕМ ДЛЯ КУРСА!
-    setTimeout(() => resolve(import('./LoginForm')), 1500);
-}));
- */
