--- src/features/Example AuthByUsername/ui/LoginForm/LoginForm.module.scss
@@ -1,14 +1,0 @@
-.LoginForm {
-    display: flex;
-    flex-direction: column;
-    width: 400px;
-}
-
-.input {
-    margin-top: 10px;
-}
-
-.loginBtn {
-    margin-top: 15px;
-    margin-left: auto;
-}
