--- src/features/Example AuthByUsername/ui/LoginForm/LoginForm.sstories.tsx
@@ -1,34 +1,0 @@
-/* import React from 'react';
-import { ComponentMeta, ComponentStory } from '@storybook/react';
-import { StoreDecorator } from 'shared/config/storybook/StoreDecorator/StoreDecorator';
-import { PreloadedState } from '@reduxjs/toolkit';
-import LoginForm from './LoginForm';
-
-export default {
-    title: 'features/LoginForm',
-    component: LoginForm,
-    argTypes: {
-        backgroundColor: { control: 'color' },
-    },
-} as ComponentMeta<typeof LoginForm>;
-
-const Template: ComponentStory<typeof LoginForm> = (args) => <LoginForm {...args} />;
-
-export const Primary = Template.bind({});
-Primary.args = {};
-Primary.decorators = [StoreDecorator({
-    loginForm: { username: '123', password: 'asd' },
-})];
-
-export const withError = Template.bind({});
-withError.args = {};
-withError.decorators = [StoreDecorator({
-    loginForm: { username: '123', password: 'asd', error: 'ERROR' },
-})];
-
-export const Loading = Template.bind({});
-Loading.args = {};
-Loading.decorators = [StoreDecorator({
-    loginForm: { isLoading: true },
-})];
- */
