--- src/features/Example AuthByUsername/ui/LoginModal/LoginModal.tsx
@@ -1,25 +1,0 @@
-/* import { Modal } from "shared/ui/Modal/Modal"
-import { classNames } from "shared/lib/classNames/classNames"
-import { Suspense } from "react"
-import { Loader } from "shared/ui/Loader/Loader"
-import { LoginFormAsync } from "../LoginForm/LoginForm.async"
-
-interface LoginModalProps {
-    className?: string
-    isOpen: boolean
-    onClose: () => void
-}
-
-export const LoginModal = ({ className, isOpen, onClose }: LoginModalProps) => (
-    <Modal
-        className={classNames("", {}, [className])}
-        isOpen={isOpen}
-        onClose={onClose}
-        lazy
-    >
-        <Suspense fallback={<Loader />}>
-            <LoginFormAsync onSuccess={onClose} />
-        </Suspense>
-    </Modal>
-)
- */
