+++ src/features/FilterProducts/index.ts
@@ -1,0 +1,1 @@
+export { FilterProducts } from "./ui/ProductFilters"
