+++ src/features/FilterProducts/ui/CheckboxGroup/CheckboxGroup.module.scss
