+++ src/features/FilterProducts/ui/CheckboxGroup/CheckboxGroup.tsx
