+++ src/features/FilterProducts/ui/PriceFilter/PriceFilter.module.scss
