+++ src/features/FilterProducts/ui/PriceFilter/PriceFilter.tsx
