--- src/features/ProductFilters/index.ts
@@ -1,3 +1,0 @@
-import { ProductFilters } from "./ui/ProductFilters"
-
-export { ProductFilters }
