--- src/widgets/FiltersModalSlider/index.ts
@@ -1,3 +1,0 @@
-import { FiltersModalSlider } from "./ui/FiltersModalSlider"
-
-export { FiltersModalSlider }
